chore: rename bottomsheets folder
